+++ noline/src/sync.rs
@@ -6,6 +6,7 @@
 use ::core::marker::PhantomData;
 
 use crate::error::Error;
+use crate::history::History;
 use crate::line_buffer::{Buffer, LineBuffer};
 
 use crate::core::{Initializer, InitializerResult, Line};
@@ -32,15 +33,17 @@ pub trait Write {
 }
 
 /// Line editor for synchronous IO
-pub struct Editor<B: Buffer, IO: Read + Write> {
+pub struct Editor<B: Buffer, H: History, IO: Read + Write> {
     buffer: LineBuffer<B>,
     terminal: Terminal,
+    history: H,
     _marker: PhantomData<IO>,
 }
 
-impl<B, IO, RE, WE> Editor<B, IO>
+impl<B, H, IO, RE, WE> Editor<B, H, IO>
 where
     B: Buffer,
+    H: History,
     IO: Read<Error = RE> + Write<Error = WE>,
 {
     /// Create and initialize line editor
@@ -64,6 +67,7 @@ where
         Ok(Self {
             buffer: LineBuffer::new(),
             terminal,
+            history: H::default(),
             _marker: PhantomData,
         })
     }
@@ -92,7 +96,12 @@ where
         prompt: &'b str,
         io: &mut IO,
     ) -> Result<&'b str, Error<RE, WE>> {
-        let mut line = Line::new(prompt, &mut self.buffer, &mut self.terminal);
+        let mut line = Line::new(
+            prompt,
+            &mut self.buffer,
+            &mut self.terminal,
+            &mut self.history,
+        );
         Self::handle_output(line.reset(), io)?;
 
         loop {
@@ -104,10 +113,22 @@ where
 
         Ok(self.buffer.as_str())
     }
+
+    /// Load history from iterator
+    pub fn load_history<'a>(&mut self, entries: impl Iterator<Item = &'a str>) -> usize {
+        self.history.load_entries(entries)
+    }
+
+    /// Get history as iterator over circular slices
+    pub fn get_history<'a>(&'a self) -> impl Iterator<Item = CircularSlice<'a>> {
+        get_history_entries(&self.history)
+    }
 }
 
 #[cfg(test)]
 mod tests {
+    use crate::history::NoHistory;
+
     use super::*;
     use ::std::string::ToString;
     use ::std::{thread, vec::Vec};
@@ -164,7 +185,7 @@ mod tests {
         let mut io = IO::new(input_rx, output_tx);
 
         let handle = thread::spawn(move || {
-            let mut editor: Editor<Vec<u8>, _> = Editor::new(&mut io).unwrap();
+            let mut editor: Editor<Vec<u8>, NoHistory, _> = Editor::new(&mut io).unwrap();
 
             if let Ok(s) = editor.readline("> ", &mut io) {
                 Some(s.to_string())
@@ -285,6 +306,7 @@ pub mod std {
 
         use crossbeam::channel::{unbounded, Receiver, Sender};
 
+        use crate::history::NoHistory;
         use crate::sync::Editor;
         use crate::testlib::{test_cases, test_editor_with_case, MockTerminal};
         use ::std::io::Read as IoRead;
@@ -394,7 +416,7 @@ pub mod std {
                     |(stdin, stdout), string_tx| {
                         thread::spawn(move || {
                             let mut io = IO::new(stdin, stdout);
-                            let mut editor = Editor::<Vec<u8>, _>::new(&mut io).unwrap();
+                            let mut editor = Editor::<Vec<u8>, NoHistory, _>::new(&mut io).unwrap();
 
                             while let Ok(s) = editor.readline(prompt, &mut io) {
                                 string_tx.send(s.to_string()).unwrap();
@@ -507,6 +529,7 @@ pub mod embedded {
 
         use crossbeam::channel::{Receiver, Sender, TryRecvError};
 
+        use crate::history::NoHistory;
         use crate::line_buffer::StaticBuffer;
         use crate::sync::Editor;
         use crate::testlib::test_editor_with_case;
@@ -578,7 +601,8 @@ pub mod embedded {
                     |serial, string_tx| {
                         thread::spawn(move || {
                             let mut io = IO::new(serial);
-                            let mut editor = Editor::<StaticBuffer<100>, _>::new(&mut io).unwrap();
+                            let mut editor =
+                                Editor::<StaticBuffer<100>, NoHistory, _>::new(&mut io).unwrap();
 
                             while let Ok(s) = editor.readline(prompt, &mut io) {
                                 string_tx.send(s.to_string()).unwrap();